Messing around
This is all pulled apart but is starting to be in the right direction. Still needs a ton of work to get the object passed back to be usefull in the server side controller. Im only pushing so you can see what I was doing. There are some notes sprinkled in.